+++ example/lib/views/rewarded_example.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:adrop_ads_flutter/adrop_ads_flutter.dart';
 import 'package:adrop_ads_flutter_example/test_unit_ids.dart';
 import 'package:flutter/material.dart';
@@ -19,10 +21,15 @@ class _RewardedExampleState extends State<RewardedExample> {
 
   bool disabledReset() => !(errorCode != null || isShown);
 
+  String unit() {
+    // Use your actual rewarded ad unit IDs here
+    return Platform.isAndroid ? testUnitIdRewardedAd : testUnitIdRewardedAd;
+  }
+
   @override
   void initState() {
     super.initState();
-    reset(testUnitIdRewardedAd);
+    reset(unit());
   }
 
   void reset(String unitId) {
@@ -105,7 +112,7 @@ class _RewardedExampleState extends State<RewardedExample> {
                         onPressed: disabledReset()
                             ? null
                             : () {
-                                reset(testUnitIdRewardedAd);
+                                reset(unit());
                               },
                         child: const Text('reset (test ad)')),
                     const Text(